Begin modelling architectural skeleton
What:
This commit adds the basic architectural structure for some of the application's core classes and abstractions.
It consists of only declarations - no definitions or implementation. It is incomplete and therefore requires more of the same type of commits, as the architecture continues to be planned and built.

Why:
To get the ball rolling producing the skeleton for the application.