Add app minimise to tray for Dashboard form
What:
Added the abillity to minimise the application to the system tray and interact with its basic open/quit functionality from a context menu.

Why:
Although this is taking a form-oriented approach to system tray minimisation, rather than a tray icon approach, it's suitable for the use case currently.
It can and will either be adapted to other forms as needed (there may only be 1 or two other forms), or, refactored to the
notify icon approach as described by Michael Sorens at Redgate, here:
https://www.red-gate.com/simple-talk/dotnet/.net-framework/creating-tray-applications-in-.net-a-practical-guide/#ninth